Extract API-encryption key helpers into yaml/api_encryption.py
Third PR in the helpers/yaml split arc (follows #796 + #798).
Moves generate_api_encryption_key and rewrite_api_encryption_key
byte-for-byte from yaml/__init__.py into the new
yaml/api_encryption.py submodule.

api_encryption.py imports the scalar helpers it needs (_quote,
_strip_yaml_quotes, rewrite_yaml_scalar) from the parent
package — same pattern as #798's substitution.py.

__init__.py drops the now-unused top-level imports of base64
and secrets (both only used by generate_api_encryption_key,
which moved with the function).

Re-exports added at the bottom of __init__.py in PEP 484
redundant-alias form so external callers
(controllers/devices/mutations_clone.py, tests) keep using
"from .helpers.yaml import generate_api_encryption_key"
unchanged.

__init__.py: 1015 -> 975 lines (-40). api_encryption.py: 48 lines.

All 3404 non-e2e tests pass; ruff + mypy clean.
